Update home page and navigation bar
