+++ package.json
@@ -33,6 +33,7 @@
     "react-router-dom": "^6.26.0",
     "react-scripts": "5.0.1",
     "supertest": "^7.0.0",
+    "typeface-inter": "^3.18.1",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
