+++ src/js/components/App.js
@@ -16,104 +16,15 @@ import "@fortawesome/fontawesome-free/css/all.min.css";
 import SummaryPage from "../pages/SummaryPage";
 import Register from "../pages/Register";
 import Login from "../pages/Login";
-import { getProfile } from "../services/ProfileService"; // Import getProfile
 import { Buffer } from "buffer";
+import Navbar from "./Navbar";
 
 function App() {
-  const [isAuthenticated, setIsAuthenticated] = useState(false);
-  const [profile, setProfile] = useState(null); // Add profile state
-
-  useEffect(() => {
-    const token = localStorage.getItem("token");
-    if (token) {
-      setIsAuthenticated(true);
-      fetchProfile(); // Fetch profile data if authenticated
-    }
-  }, []);
-
-  const fetchProfile = async () => {
-    const profileData = await getProfile();
-    setProfile(profileData);
-  };
 
   return (
     <Router>
       <div className="d-flex flex-column min-vh-100">
-        {/* Navbar component for site navigation */}
-        <nav className="navbar navbar-expand-lg navbar-light bg-light">
-          <Link className="navbar-brand" to="/">
-            <img
-              src="images/logo.png"
-              alt="Search Icon"
-              width="50"
-              height="50"
-            />
-            SMARTJ
-          </Link>
-          <button
-            className="navbar-toggler"
-            type="button"
-            data-bs-toggle="collapse"
-            data-bs-target="#navbarNav"
-            aria-controls="navbarNav"
-            aria-expanded="false"
-            aria-label="Toggle navigation"
-          >
-            <span className="navbar-toggler-icon"></span>
-          </button>
-          <div className="collapse navbar-collapse" id="navbarNav">
-            <ul className="navbar-nav ms-auto">
-              {/* Navigation links */}
-              <li className="nav-item">
-                <Link className="nav-link" to="/">
-                  Home
-                </Link>
-              </li>
-              <li className="nav-item">
-                <Link className="nav-link" to="/interview-settings">
-                  Interview Practice
-                </Link>
-              </li>
-              <li className="nav-item">
-                <Link className="nav-link" to="/job-finder">
-                  Job Finder
-                </Link>
-              </li>
-              <li className="nav-item">
-                <Link className="nav-link" to="/contact-us">
-                  Contact Us
-                </Link>
-              </li>
-              {isAuthenticated ? (
-                <>
-                  <li className="nav-item">
-                    <Link className="nav-link" to="/my-profile">
-                      {profile && profile.profilePicture.data ? (
-                        <img
-                          src={`data:${
-                            profile.profilePicture.contentType
-                          };base64,${Buffer.from(
-                            profile.profilePicture.data
-                          ).toString("base64")}`}
-                          alt="Profile"
-                          className="profile-picture-nav"
-                        />
-                      ) : (
-                        "My Profile"
-                      )}
-                    </Link>
-                  </li>
-                </>
-              ) : (
-                <li className="nav-item">
-                  <Link className="nav-link" to="/login">
-                    Login
-                  </Link>
-                </li>
-              )}
-            </ul>
-          </div>
-        </nav>
+        <Navbar />
 
         {/* Main content area that renders the routes */}
         <div className="flex-grow-1">
