+++ src/js/pages/Home.js
@@ -1,72 +1,75 @@
-import React from "react";
+import React, {useEffect, useState} from "react";
 import { Link } from "react-router-dom";
 import "../../styles/pages/Home.css";
 
 function Home() {
+  const [isAuthenticated, setIsAuthenticated] = useState(false);
+
+  useEffect(() => {
+    const token = localStorage.getItem("token");
+    if (token) {
+      setIsAuthenticated(true);
+    }
+  }, []);
+
+
   return (
-    <div className="container text-center mt-5">
-      {/* Main heading for the homepage with logo */}
-      <h1 className="display-4">
-        <img src="images/logo.png" alt="Search Icon" width="50" height="50" />
-        SMARTJ
-      </h1>
+      <div className="container-xxl text-center mt-5">
 
-      {/* Introduction paragraph explaining the purpose of the application */}
-      <p className="lead">
-        We empower tech job seekers by offering interactive interview practice
-        sessions with customizable questions, recording features, and feedback
-        for improvement. We also streamline job searching by providing curated
-        links to tech job listings, saving time and enhancing your job search
-        experience.
-      </p>
+          <div className="landing">
+              <div className="landing-heading">
+                  {/* Main heading for the homepage*/}
+                  <h1 className="landing-primary-text">
+                      SMARTJ
+                  </h1>
 
-      {/* Container for the feature cards */}
-      <div className="row mt-5">
-        {/* Card for Interview Practice */}
-        <div className="col-md-6">
-          <div className="card mb-4 shadow-sm">
-            <Link
-              to="/interview-settings"
-              className="text-decoration-none text-dark d-block"
-            >
-              {/* Image and title for Interview Practice feature */}
-              <img
-                src="images/interview_practice.png"
-                width="200"
-                height="400"
-                className="card-img-top"
-                alt="Interview Practice"
-              />
-              <div className="card-body">
-                <h5 className="card-title">Interview Practice</h5>
+                  {/* Introduction paragraph explaining the purpose of the application */}
+                  <p className="landing-subtext">
+                      Helping Active Job Seekers In Tech Prepare For <span className="gradient-text">Interviews</span>
+                  </p>
+
+                  {/* Buttons for signing up or logging in or navigation*/}
+                  {isAuthenticated ? (
+                      <div>
+                          <div className="landing-options">
+                              <div>
+                                  <Link to="/interview-settings" className="primary-btn">
+                                      Start Interview Practice
+                                  </Link>
+                              </div>
+                              <div>
+                                  <Link to="/login" className="primary-btn">
+                                      Go To Job Finder
+                                  </Link>
+                              </div>
+                          </div>
+                          <p className="landing-note-text">Jump into a new practice session or view existing jobs</p>
+                      </div>
+                  ) : (
+                      <div>
+                          <div className="landing-options">
+                              <div>
+                                  <Link to="/register" className="primary-btn">
+                                      Sign Up For Free
+                                  </Link>
+                              </div>
+                              <div>
+                                  <Link to="/login" className="secondary-btn">
+                                      Use Existing Account
+                                  </Link>
+                              </div>
+                          </div>
+                          <p className="landing-note-text">Sign Up and get Full Access to all SMARTJ tools</p>
+                      </div>
+                  )}
               </div>
-            </Link>
+              <img className="landing-image" src="images/landing-image.svg" alt="Landing Page Image"/>
           </div>
-        </div>
 
-        {/* Card for Job Finder */}
-        <div className="col-md-6">
-          <div className="card mb-4 shadow-sm">
-            <Link
-              to="/job-finder"
-              className="text-decoration-none text-dark d-block"
-            >
-              {/* Image and title for Job Finder feature */}
-              <img
-                src="images/job_finder.png"
-                width="200"
-                height="400"
-                className="card-img-top"
-                alt="Job Finder"
-              />
-              <div className="card-body">
-                <h5 className="card-title">Job Finder</h5>
-              </div>
-            </Link>
-          </div>
-        </div>
+
+          <div className="placeholder-card" style={{marginTop: 60}}/>
+          <div className="placeholder-card" style={{marginTop: 40, marginLeft: 150, marginBottom: 50}}/>
       </div>
-    </div>
   );
 }
