+++ src/styles/components/App.css
@@ -2,86 +2,13 @@
   text-align: center;
 }
 
-.App-logo {
-  height: 40vmin;
-  pointer-events: none;
-}
-
-@media (prefers-reduced-motion: no-preference) {
-  .App-logo {
-    animation: App-logo-spin infinite 20s linear;
-  }
-}
-
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
-}
-
-.App-link {
-  color: #61dafb;
-}
-
-.profile-picture {
-  width: 100px; /* Adjust the size as needed */
-  height: 100px; /* Adjust the size as needed */
-  border-radius: 50%; /* Makes the image round */
-  object-fit: cover; /* Ensures the image covers the container */
-  display: block;
-  margin: 0 auto;
-}
-
-.profile-picture-nav {
-  width: 30px;
-  height: 30px;
-  border-radius: 50%;
-  object-fit: cover;
-}
-
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-}
-
-/* Center the nav links when the menu is collapsed */
-@media (max-width: 991.98px) {
-  .navbar-collapse {
-    text-align: center;
-  }
-
-  .navbar-nav {
-    flex-direction: column;
-    width: 100%;
-  }
-
-  .nav-item {
-    width: 100%;
-  }
-
-  .nav-link {
-    width: 100%;
-    text-align: center;
-    padding: 10px 0;
-  }
-}
-
 /* FOOTER */
 .footer {
-  background-color: #67a9d2;
+  background-color: #DDDBFF;
 }
 
 .copyright-section {
   color: black;
-  background-color: #c7e2fd;
+  background-color: #aDaBfF;
 
 }
