+++ src/styles/components/Navbar.css
@@ -1,0 +1,71 @@
+.navbar {
+    border-bottom: 1px solid #DDDBFF;
+}
+
+.nav-text {
+    font-family: 'Inter', sans-serif;
+    font-weight: 900;
+    font-style: italic;
+    font-size: 32px;
+    color: #212529;
+}
+
+.profile-picture {
+    width: 100px; /* Adjust the size as needed */
+    height: 100px; /* Adjust the size as needed */
+    border-radius: 50%; /* Makes the image round */
+    object-fit: cover; /* Ensures the image covers the container */
+    display: block;
+    margin: 0 auto;
+}
+
+.profile-picture-nav {
+    width: 40px;
+    border-radius: 50%;
+    object-fit: cover;
+    margin-right: 1rem;
+}
+
+.navbar-toggle {
+    border-radius: 15px;
+    background-color: #DDDBFF;
+    border: none;
+    filter: drop-shadow(0px 4px 2px rgba(0,0,0,0.45));
+    margin-left: 1rem;
+    min-height: 40px;
+    display: none;
+}
+
+
+/* Center the nav links when the menu is collapsed */
+@media (max-width: 991.98px) {
+    .navbar-collapse {
+        text-align: center;
+    }
+
+    .navbar-nav {
+        flex-direction: column;
+        width: 100%;
+    }
+
+    .nav-item {
+        width: 100%;
+    }
+
+    .nav-link {
+        width: 100%;
+        text-align: center;
+        padding: 10px 0;
+    }
+
+    .navbar-toggle {
+        display: block;
+    }
+}
+
+@media (min-width: 991.98px) {
+    .nav-text {
+        font-size: 3.5rem;
+        margin-left: 2rem;
+    }
+}
